HHH-9475 Cannot mix @MapKey with @Convert - Moved classes for reuse in other test case
