+++ hibernate-core/src/test/java/org/hibernate/test/converter/map/CustomColorTypeConverter.java
@@ -1,0 +1,20 @@
+package org.hibernate.test.converter.map;
+
+import javax.persistence.AttributeConverter;
+import javax.persistence.Converter;
+
+/**
+ * @author Janario Oliveira
+ */
+@Converter
+public class CustomColorTypeConverter implements AttributeConverter<ColorType, String> {
+	@Override
+	public String convertToDatabaseColumn(ColorType attribute) {
+		return attribute == null ? null : "COLOR-" + attribute.toExternalForm();
+	}
+
+	@Override
+	public ColorType convertToEntityAttribute(String dbData) {
+		return dbData == null ? null : ColorType.fromExternalForm( dbData.substring( 6 ) );
+	}
+}
