+++ hibernate-core/src/test/java/org/hibernate/test/converter/map/EnumMapKey.java
@@ -1,0 +1,9 @@
+package org.hibernate.test.converter.map;
+
+/**
+ * @author Janario Oliveira
+ */
+public enum EnumMapKey {
+	VALUE_1,
+	VALUE_2
+}
