+++ hibernate-core/src/test/java/org/hibernate/test/converter/map/ImplicitEnumMapKey.java
@@ -1,0 +1,9 @@
+package org.hibernate.test.converter.map;
+
+/**
+ * @author Janario Oliveira
+ */
+public enum ImplicitEnumMapKey {
+	VALUE_1,
+	VALUE_2
+}
